# Conflicts:
#	Back/demo/src/main/java/com/example/demo/common/Constants.java
#	Back/demo/src/main/java/com/example/demo/controller/AteController.java
#	Back/demo/src/main/java/com/example/demo/controller/DishController.java
#	Back/demo/src/main/java/com/example/demo/controller/MainController.java
#	Back/demo/src/main/resources/mappers/AteDao.xml

+++ Back/demo/src/main/java/com/example/demo/controller/MemberController.java
@@ -37,11 +37,11 @@ import lombok.extern.slf4j.Slf4j;
 public class MemberController {
 
 	private final MemberService memberService;
-	
+
 	private final PointService pointService;
-	
+
 	private final AteService ateService;
-	
+
 	private final DishService dishService;
 
 	// 회원 아이디 또는 회원 식별자로 회원 조회
@@ -62,17 +62,16 @@ public class MemberController {
 
 		// 회원 아이디 중복 체크 true or false 사용
 		if (result == null) {
-			log.info("[/api/member/search/{memberID}] [searchMember 성공] [{}]",memberID);
+			log.info("[/api/member/search/{memberID}] [searchMember 성공] [{}]", memberID);
 			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 		} else {
 			log.warn("[/api/member/search/{memberID}] [searchMember 실패] [{}]", memberID);
-			
+
 			result.setMemberPW("");
-			
+
 			return new ResponseEntity<>(result, HttpStatus.OK);
 		}
 	}
-
 
 	// 회원 삭제 URI
 	@DeleteMapping("/delete/{mnum}")
@@ -82,15 +81,14 @@ public class MemberController {
 
 		// 추후 주석 해제 예정
 
-		/*
-		 * long getNumber = (long) request.getAttribute("GetNumber");
-		 * 
-		 * if (mnum != getNumber) {
-		 * 
-		 * log.warn("[/api/member/delete/{mnum}] [removeAdmin 실패] [{}]", mnum);
-		 * 
-		 * return new ResponseEntity<>(HttpStatus.FORBIDDEN); }
-		 */
+		long getNumber = (long) request.getAttribute("GetNumber");
+
+		if (mnum != getNumber) {
+
+			log.warn("[/api/member/delete/{mnum}] [removeAdmin 실패] [{}]", mnum);
+
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
 
 		// 정상적으로 삭제됐으면 anum 리턴, 아니면 0
 		Long result = memberService.remove(mnum);
@@ -100,7 +98,7 @@ public class MemberController {
 			return new ResponseEntity<>(true, HttpStatus.OK);
 		} else {
 			log.warn("[/api/member/delete/{mnum}] [removeMember 실패] [{}]", mnum);
-			return new ResponseEntity<>(false,HttpStatus.BAD_REQUEST);
+			return new ResponseEntity<>(false, HttpStatus.BAD_REQUEST);
 		}
 
 	}
@@ -113,15 +111,14 @@ public class MemberController {
 
 		// 추후 주석 해제 예정
 
-		
-//		  long getNumber = (long) request.getAttribute("GetNumber");
-//		  
-//		  if (memberDTO.getMnum() != getNumber) {
-//		  
-//		  log.warn("[/api/member/update] [updateAdmin 실패] [{}]", memberDTO.getMnum());
-//		  
-//		  return new ResponseEntity<>(HttpStatus.FORBIDDEN); }
-		 
+		long getNumber = (long) request.getAttribute("GetNumber");
+
+		if (memberDTO.getMnum() != getNumber) {
+
+			log.warn("[/api/member/update] [updateAdmin 실패] [{}]", memberDTO.getMnum());
+
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
 
 		Long result = memberService.update(memberDTO);
 
@@ -130,170 +127,167 @@ public class MemberController {
 			return new ResponseEntity<>(true, HttpStatus.OK);
 		} else {
 			log.warn("[/api/member/delete/{mnum}] [updateMember 실패] [{}]", memberDTO.getMemberID());
-			return new ResponseEntity<>(false,HttpStatus.BAD_REQUEST);
+			return new ResponseEntity<>(false, HttpStatus.BAD_REQUEST);
 		}
 
 	}
-	
+
 	// 포인트 조회
 	@GetMapping("/point-list/{mnum}")
-	public ResponseEntity<PageResultVO<PointDTO>> userPoint(
-			@PathVariable long mnum, PageRequestDTO pageRequestDTO){
-		
-		log.info("[/api/point-list/{mnum}] [userPoint] [{}]",mnum);
-		
+	public ResponseEntity<PageResultVO<PointDTO>> userPoint(@PathVariable long mnum, PageRequestDTO pageRequestDTO,
+			HttpServletRequest request) {
+
+		log.info("[/api/point-list/{mnum}] [userPoint] [{}]", mnum);
+
 		// 추후 주석 해제 예정
 
-		/*
-		 * long getNumber = (long) request.getAttribute("GetNumber");
-		 * 
-		 * if (mnum != getNumber) {
-		 * 
-		 * log.warn("[/api/point-list/{mnum}] [userPoint 실패] [{}]", mnum);
-		 * 
-		 * return new ResponseEntity<>(HttpStatus.FORBIDDEN); }
-		 */
-		
-		PageResultVO<PointDTO> result =  pointService.userPoint(pageRequestDTO,mnum);
-		
-		if(result == null) return new ResponseEntity<>(new PageResultVO<PointDTO>(),HttpStatus.ACCEPTED);
-		
+		long getNumber = (long) request.getAttribute("GetNumber");
+
+		if (mnum != getNumber) {
+
+			log.warn("[/api/point-list/{mnum}] [userPoint 실패] [{}]", mnum);
+
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
+
+		PageResultVO<PointDTO> result = pointService.userPoint(pageRequestDTO, mnum);
+
+		if (result == null)
+			return new ResponseEntity<>(new PageResultVO<PointDTO>(), HttpStatus.ACCEPTED);
+
 		return new ResponseEntity<>(result, HttpStatus.OK);
 	}
-	
-	//7일 이내 활동 내역
+
+	// 7일 이내 활동 내역
 	@GetMapping("/activity-history/{mnum}")
-	public ResponseEntity<PageResultVO<PointDTO>> activityHistory(
-			@PathVariable long mnum, PageRequestDTO pageRequestDTO){
-		
-		log.info("[/api/member/activity-history/{mnum}] [activityHistory] [{}]",mnum);
-		
+	public ResponseEntity<PageResultVO<PointDTO>> activityHistory(@PathVariable long mnum,
+			PageRequestDTO pageRequestDTO, HttpServletRequest request) {
+
+		log.info("[/api/member/activity-history/{mnum}] [activityHistory] [{}]", mnum);
+
 		// 추후 주석 해제 예정
 
-		/*
-		 * long getNumber = (long) request.getAttribute("GetNumber");
-		 * 
-		 * if (mnum != getNumber) {
-		 * 
-		 * log.warn("[/api/member/activity-history/{mnum}] [activityHistory 실패] [{}]",mnum);
-		 * 
-		 * return new ResponseEntity<>(HttpStatus.FORBIDDEN); }
-		 */
-		
-		PageResultVO<PointDTO> result =  pointService.userPointby7d(pageRequestDTO,mnum);
-		
-		if(result == null) return new ResponseEntity<>(new PageResultVO<PointDTO>(),HttpStatus.ACCEPTED);
-		
+		long getNumber = (long) request.getAttribute("GetNumber");
+
+		if (mnum != getNumber) {
+
+			log.warn("[/api/member/activity-history/{mnum}] [activityHistory 실패] [{}]", mnum);
+
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
+
+		PageResultVO<PointDTO> result = pointService.userPointby7d(pageRequestDTO, mnum);
+
+		if (result == null)
+			return new ResponseEntity<>(new PageResultVO<PointDTO>(), HttpStatus.ACCEPTED);
+
 		return new ResponseEntity<>(result, HttpStatus.OK);
 	}
-	
-	//좋아요 조회
+
+	// 좋아요 조회
 	@GetMapping("/like-list/{mnum}")
-	public ResponseEntity<PageResultVO<DishLikeDTO>> likeList(
-			@PathVariable long mnum, PageRequestDTO pageRequestDTO){
-		
-		log.info("[/api/member/like-list/{mnum}] [likeList] [{}]",mnum);	
-		
+	public ResponseEntity<PageResultVO<DishLikeDTO>> likeList(@PathVariable long mnum, PageRequestDTO pageRequestDTO,
+			HttpServletRequest request) {
+
+		log.info("[/api/member/like-list/{mnum}] [likeList] [{}]", mnum);
+
 		// 추후 주석 해제 예정
 
-		/*
-		 * long getNumber = (long) request.getAttribute("GetNumber");
-		 * 
-		 * if (mnum != getNumber) {
-		 * 
-		 * log.warn("[/api/member/like-list/{mnum}] [likeList 실패] [{}]",mnum);	
-		 * 
-		 * return new ResponseEntity<>(HttpStatus.FORBIDDEN); }
-		 */
-		
+		long getNumber = (long) request.getAttribute("GetNumber");
+
+		if (mnum != getNumber) {
+
+			log.warn("[/api/member/like-list/{mnum}] [likeList 실패] [{}]", mnum);
+
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
+
 		PageResultVO<DishLikeDTO> result = dishService.getLikeListbyMnum(pageRequestDTO, mnum);
-		
-		
-		
-		if(result == null) return new ResponseEntity<>(new PageResultVO<DishLikeDTO>(),HttpStatus.ACCEPTED);
-		
+
+		if (result == null)
+			return new ResponseEntity<>(new PageResultVO<DishLikeDTO>(), HttpStatus.ACCEPTED);
+
 		return new ResponseEntity<>(result, HttpStatus.OK);
 	}
-	
-	//댓글 조회
-		@GetMapping("/dishcomm-list/{mnum}")
-		public ResponseEntity<PageResultVO<DishCommDTO>> commList(
-				@PathVariable long mnum, PageRequestDTO pageRequestDTO){
-			
-			log.info("[/api/member/dishcomm-list/{mnum}] [commList] [{}]",mnum);
-			
-			// 추후 주석 해제 예정
 
-			/*
-			 * long getNumber = (long) request.getAttribute("GetNumber");
-			 * 
-			 * if (mnum != getNumber) {
-			 * 
-			 * log.warn("[/api/member/dishcomm-list/{mnum}] [commList 실패] [{}]",mnum);
-			 * 
-			 * return new ResponseEntity<>(HttpStatus.FORBIDDEN); }
-			 */
-			
-			PageResultVO<DishCommDTO> result  = dishService.getCommListbyMnum(pageRequestDTO, mnum);
-			
-			if(result == null) return new ResponseEntity<>(new PageResultVO<DishCommDTO>(),HttpStatus.ACCEPTED);
-			
-			return new ResponseEntity<>(result, HttpStatus.OK);
+	// 댓글 조회
+	@GetMapping("/dishcomm-list/{mnum}")
+	public ResponseEntity<PageResultVO<DishCommDTO>> commList(@PathVariable long mnum, PageRequestDTO pageRequestDTO,
+			HttpServletRequest request) {
+
+		log.info("[/api/member/dishcomm-list/{mnum}] [commList] [{}]", mnum);
+
+		// 추후 주석 해제 예정
+
+		long getNumber = (long) request.getAttribute("GetNumber");
+
+		if (mnum != getNumber) {
+
+			log.warn("[/api/member/dishcomm-list/{mnum}] [commList 실패] [{}]", mnum);
+
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
 		}
+
+		PageResultVO<DishCommDTO> result = dishService.getCommListbyMnum(pageRequestDTO, mnum);
+
+		if (result == null)
+			return new ResponseEntity<>(new PageResultVO<DishCommDTO>(), HttpStatus.ACCEPTED);
+
+		return new ResponseEntity<>(result, HttpStatus.OK);
+	}
+
+	// 먹음 리스트 조회(페이지 처리 안한 것)
+	@GetMapping("/getList/{mnum}")
+	public ResponseEntity<List<Ate>> getUserList(@PathVariable long mnum, HttpServletRequest request) {
+
+		log.info("[/api/getList/{mnum}] [getUserList] ] [{}]", mnum);
+
+		// 추후 주석 해제 예정
+
 		
-		//먹음 리스트 조회(페이지 처리 안한 것)
-		@GetMapping("/getList/{mnum}")
-		public ResponseEntity<List<Ate>> getUserList(@PathVariable long mnum, HttpServletRequest request) {
-			
-			log.info("[/api/getList/{mnum}] [getUserList] ] [{}]", mnum);
-			
-			
-			// 추후 주석 해제 예정
+		  long getNumber = (long) request.getAttribute("GetNumber");
+		  
+		  if (mnum != getNumber) {
+		  
+		  log.warn("[/api/getList/{mnum}] [getUserList 실패] ] [{}]", mnum);;
+		  
+		  return new ResponseEntity<>(HttpStatus.FORBIDDEN); }
+		 
 
-			/*
-			 * long getNumber = (long) request.getAttribute("GetNumber");
-			 * 
-			 * if (mnum != getNumber) {
-			 * 
-			 * log.warn("[/api/getList/{mnum}] [getUserList 실패] ] [{}]", mnum);;
-			 * 
-			 * return new ResponseEntity<>(HttpStatus.FORBIDDEN); }
-			 */	  
-			 
-			List<Ate> result = ateService.getAllList(mnum);
+		List<Ate> result = ateService.getAllList(mnum);
 
-			log.info("[/api/getList/{mnum}] [getUserList 성공] [{}]", mnum);
-			
-			return new ResponseEntity<>(result, HttpStatus.OK);
-		}
+		log.info("[/api/getList/{mnum}] [getUserList 성공] [{}]", mnum);
+
+		return new ResponseEntity<>(result, HttpStatus.OK);
+	}
+
+	// 먹음 리스트 조회(페이지 처리 한 것)
+	@GetMapping("/ate-list/{mnum}")
+	public ResponseEntity<PageResultVO<AteDTO>> getAteUserList2(@PathVariable long mnum, HttpServletRequest request,
+			PageRequestDTO pageRequestDTO) {
+
+		log.info("[/api/member/ate-list/{mnum}] [getAteUserList2] [{}]", mnum);
+
+		// 추후 주석 해제 예정
+
 		
-		//먹음 리스트 조회(페이지 처리 한 것)
-		@GetMapping("/ate-list/{mnum}")
-		public ResponseEntity<PageResultVO<AteDTO>> getAteUserList2(@PathVariable long mnum, HttpServletRequest request,
-				PageRequestDTO pageRequestDTO) {
-			
-			log.info("[/api/member/ate-list/{mnum}] [getAteUserList2] [{}]", mnum);
-			
-			// 추후 주석 해제 예정
+		  long getNumber = (long) request.getAttribute("GetNumber");
+		  
+		  if (mnum != getNumber) {
+		  
+		  log.warn("[/api/member/ate-list/{mnum}] [getAteUserList2 실패] ] [{}]", mnum);
+		  
+		  return new ResponseEntity<>(HttpStatus.FORBIDDEN); }
+		 
 
-			/*
-			 * long getNumber = (long) request.getAttribute("GetNumber");
-			 * 
-			 * if (mnum != getNumber) {
-			 * 
-			 * log.warn("[/api/member/ate-list/{mnum}] [getAteUserList2 실패] ] [{}]", mnum);
-			 * 
-			 * return new ResponseEntity<>(HttpStatus.FORBIDDEN); }
-			 */
-			
-			PageResultVO<AteDTO> result = ateService.getUserAteList(pageRequestDTO,mnum);
+		PageResultVO<AteDTO> result = ateService.getUserAteList(pageRequestDTO, mnum);
 
-			if(result == null) return new ResponseEntity<>(new PageResultVO<AteDTO>(),HttpStatus.ACCEPTED);
-			
-			log.info("[/api/member/ate-list/{mnum}] [getAteUserList2 성공] [{}]", mnum);
-			
-			return new ResponseEntity<>(result, HttpStatus.OK);
-		}
-	
+		if (result == null)
+			return new ResponseEntity<>(new PageResultVO<AteDTO>(), HttpStatus.ACCEPTED);
+
+		log.info("[/api/member/ate-list/{mnum}] [getAteUserList2 성공] [{}]", mnum);
+
+		return new ResponseEntity<>(result, HttpStatus.OK);
+	}
 
 }
